microservices/auth_service/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
requests
passlib[bcrypt]